Consolidate redundant test files
Merge reader_bounds_test.go into bad_data_test.go, and merge
internal/decoder/example_kind_test.go into
internal/decoder/kind_test.go.

This reduces the number of redundant test files and keeps the test suite
focused and clean.
